Idealized Variance Fractions

Compute idealized variance fractions removing CountNoise or other variables

Usage

idealized(x, remove = "CountNoise")

Arguments

x

matrix or data.frame of variance fracdtions

remove

colnames of x to remove from the variance fractions

Details

Variance partitioning analysis includes all components of variance in the denominator when computing the variance fractions. This function recomputes the variance fractions in idealized case where CountNoise or any set of variances given in the remove argument is zero.

Examples

# Simulate counts
set.seed(1)
countMatrix <- matrix(rnbinom(n=100000, mu=20, size=3), ncol=10)
rownames(countMatrix) <- paste0("gene_", seq(nrow(countMatrix)))
colnames(countMatrix) <- paste0("sample_", seq(ncol(countMatrix)))

condition <- factor(rep(1:2, each=5))

# DESeq2 model #
library(DESeq2)
dds <- DESeqDataSetFromMatrix(countMatrix, 
  DataFrame(condition), 
  ~ condition)
#> converting counts to integer mode
dds <- DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- results(dds)

# Variance partition analysis
vp1 <- varpart(dds)

# Compute variance fractions in the idealized
#  case where there is no count noise
head(idealized(vp1))
#>           condition Residuals
#> gene_1 0.3545538734 0.6454461
#> gene_2 0.0001032731 0.9998967
#> gene_3 0.4544569666 0.5455430
#> gene_4 0.0264197616 0.9735802
#> gene_5 0.1288757222 0.8711243
#> gene_6 0.0010682706 0.9989317