Residuals from model fit

Extract residuals for each gene from model fit with fitVarPartModel()

# S4 method for VarParFitList
residuals(object, ...)

Arguments

object
object produced by fitVarPartModel()

...
other arguments.

Value

Residuals extracted from model fits stored in object

Details

If model is fit with missing data, residuals returns NA for entries that were missing in the original data

Examples

# load library
# library(variancePartition)

# Intialize parallel backend with 4 cores
library(BiocParallel)
register(SnowParam(4))

# load simulated data:
# geneExpr: matrix of gene expression values
# info: information/metadata about each sample
data(varPartData)

# Specify variables to consider
# Age is continuous so we model it as a fixed effect
# Individual and Tissue are both categorical, so we model them as random effects
form <- ~ Age + (1|Individual) + (1|Tissue) 

# Fit model
modelFit <- fitVarPartModel( geneExpr, form, info )

#> Memory usage to store result: >8.9 Mb

#> Dividing work into 10 chunks...

#> 
#> Total:10 s

# Extract residuals of model fit
res <- residuals( modelFit )