Transform RNA-Seq Data Ready for Linear Mixed Modelling with dream()

Transform count data to log2-counts per million (logCPM), estimate the mean-variance relationship and use this to compute appropriate observation-level weights. The data are then ready for linear mixed modelling with dream(). This method is the same as limma::voom(), except that it allows random effects in the formula

voomWithDreamWeights(
  counts,
  formula,
  data,
  lib.size = NULL,
  normalize.method = "none",
  span = 0.5,
  weights = NULL,
  plot = FALSE,
  save.plot = FALSE,
  quiet = FALSE,
  BPPARAM = bpparam(),
  ...
)

Arguments

counts
a numeric matrix containing raw counts, or an ExpressionSet containing raw counts, or a DGEList object. Counts must be non-negative and NAs are not permitted.

formula
specifies variables for the linear (mixed) model. Must only specify covariates, since the rows of exprObj are automatically used a a response. e.g.: ~ a + b + (1|c) Formulas with only fixed effects also work, and lmFit() followed by contrasts.fit() are run.

data
data.frame with columns corresponding to formula

lib.size
numeric vector containing total library sizes for each sample. Defaults to the normalized (effective) library sizes in counts if counts is a DGEList or to the columnwise count totals if counts is a matrix.

normalize.method
the microarray-style normalization method to be applied to the logCPM values (if any). Choices are as for the method argument of normalizeBetweenArrays when the data is single-channel. Any normalization factors found in counts will still be used even if normalize.method="none".

span
width of the lowess smoothing window as a proportion.

weights
Can be a numeric matrix of individual weights of same dimensions as the counts, or a numeric vector of sample weights with length equal to ncol(counts)

plot
logical, should a plot of the mean-variance trend be displayed?

save.plot
logical, should the coordinates and line of the plot be saved in the output?

quiet
suppress message, default FALSE

BPPARAM
parameters for parallel evaluation

...
other arguments are passed to lmer.

Value

An EList object just like the result of limma::voom()

Details

Adapted from vomm() in limma v3.40.2

See also

Examples

# library(variancePartition)
library(edgeR)
library(BiocParallel)

data(varPartDEdata)

# normalize RNA-seq counts
dge = DGEList(counts = countMatrix)
dge = calcNormFactors(dge)

# specify formula with random effect for Individual
form <- ~ Disease + (1|Individual) 

# compute observation weights
vobj = voomWithDreamWeights( dge[1:20,], form, metadata)

#> Memory usage to store result: >49.7 Kb

#> Dividing work into 1 chunks...

#> 
#> Total:4 s

# fit dream model 
res = dream( vobj, form, metadata)

#> Dividing work into 1 chunks...

#> 
#> Total:4 s

# extract results
topTable(res, coef="Disease1")

#>                                         logFC  AveExpr        t      P.Value
#> ENST00000456159.1 gene=MET          1.0182945 2.458926 6.707820 9.637382e-07
#> ENST00000418210.2 gene=TMEM64       1.0375652 4.715367 8.047821 4.039375e-06
#> ENST00000555834.1 gene=RPS6KL1      0.9355651 5.272063 6.029943 4.544895e-06
#> ENST00000248564.5 gene=GNG11        1.0543333 4.511181 5.184523 3.373837e-05
#> ENST00000570099.1 gene=YPEL3        0.9654350 2.063331 7.009797 3.876342e-05
#> ENST00000317802.7 gene=TSPYL6       1.0257641 4.321189 4.662013 9.772471e-04
#> ENST00000482796.1 gene=RP11-101E3.5 0.9257951 1.533654 3.753019 1.099478e-03
#> ENST00000267622.4 gene=TRIP11       0.6226163 5.942593 3.672330 1.336570e-03
#> ENST00000337859.6 gene=ZC3H15       0.8379527 4.169523 3.552296 1.785163e-03
#> ENST00000598661.1 gene=AL136376.1   0.8931155 1.231360 4.229887 1.792281e-03
#>                                        adj.P.Val    z.std
#> ENST00000456159.1 gene=MET          1.927476e-05 4.898902
#> ENST00000418210.2 gene=TMEM64       3.029930e-05 4.609346
#> ENST00000555834.1 gene=RPS6KL1      3.029930e-05 4.584769
#> ENST00000248564.5 gene=GNG11        1.550537e-04 4.146644
#> ENST00000570099.1 gene=YPEL3        1.550537e-04 4.114729
#> ENST00000317802.7 gene=TSPYL6       3.141365e-03 3.296997
#> ENST00000482796.1 gene=RP11-101E3.5 3.141365e-03 3.263751
#> ENST00000267622.4 gene=TRIP11       3.341425e-03 3.208010
#> ENST00000337859.6 gene=ZC3H15       3.385222e-03 3.123826
#> ENST00000598661.1 gene=AL136376.1   3.385222e-03 3.122654