Empirical Bayes Fully Moderated t-statistics

Empirical Bayes Fully Moderated t-statistics from linear mixed model fit with dream

Usage

eBayesFMT(fit, data, Individual, method = c("VC", "WS"))

Arguments

fit

model fit returned by dream of class MArrayLM2

data

data.frame with columns corresponding to formula

Individual

string referring to column in data.frame indicating which individual the repeated measures come from

method

Use either variance components ('VC') or Welch-Satterthwaite ('WS')

Details

Applies empirical Bayes method of, Yu, et al. (2019) for linear mixed models. This method applies a prior and shrinkage to 1) the residual variance and 2) the variance component estimates for 'Individual'. 'Individual' refers to the variable indicating which individual the repeated measures come from. This method then combines posterior values from (1) and (2) to approximate the degrees of freedom for the t-statistic.

Yu, L., Zhang, J., Brock, G. et al. Fully moderated t-statistic in linear modeling of mixed effects for differential expression analysis. BMC Bioinformatics 20, 675 (2019). https://doi.org/10.1186/s12859-019-3248-9

Examples

# load library
# library(variancePartition)
library(BiocParallel)

# load simulated data:
# geneExpr: matrix of gene expression values
# info: information/metadata about each sample
data(varPartData)

form <- ~ Batch + (1|Individual) + (1|Tissue) 

# Fit linear mixed model for each gene
# run on just 10 genes for time
fit = dream( geneExpr[1:10,], form, info)
#> Dividing work into 1 chunks...
#> 
#> Total:0.6 s

# view top genes using standard t-statistics
topTable( fit )
#> Removing intercept from test coefficients
#>             Batch2       Batch3      Batch4     AveExpr          F   P.Value
#> gene8   0.02567263 -0.129563777  0.67499421   0.9171386 1.96561914 0.1267816
#> gene6  -0.67049689 -1.034776107 -1.04729128  -3.1554790 1.93232542 0.1319909
#> gene3   0.08200976 -0.172582199 -0.49673001   0.1702122 1.44637527 0.2364513
#> gene1  -0.46092425 -0.453624935 -0.05771812 -10.4664549 1.03262351 0.3833791
#> gene7  -0.31606553  0.149452306 -0.03562447  -4.3799381 0.74258476 0.5301575
#> gene2  -0.49420745 -0.382210977 -0.40013018  -1.1281610 0.62797976 0.5992798
#> gene5   0.26795357  0.006060832  0.27368297   4.7187640 0.32958690 0.8039606
#> gene9  -0.35301416 -0.096570091 -0.12646906  -2.3079042 0.27897593 0.8403964
#> gene10  0.11578714 -0.098856729  0.14565869  -2.3673775 0.27305129 0.8446423
#> gene4   0.11531023  0.122855760  0.13721184  -4.5359748 0.05418364 0.9832565
#>        adj.P.Val      F.std
#> gene8  0.6599547 1.90225556
#> gene6  0.6599547 1.87125160
#> gene3  0.7881710 1.41413220
#> gene1  0.9384914 1.01798899
#> gene7  0.9384914 0.73636776
#> gene2  0.9384914 0.62417724
#> gene5  0.9384914 0.32960191
#> gene9  0.9384914 0.27928007
#> gene10 0.9384914 0.27338242
#> gene4  0.9832565 0.05449604

# Compute moderated t-statistics from Yu, et al, 2019
fiteb = variancePartition:::eBayesFMT(fit, info, 'Individual')
topTable( fiteb )
#> Removing intercept from test coefficients
#> topTableF is obsolete and will be removed in a future version of limma. Please considering using topTable instead.
#>             Batch2       Batch3      Batch4     AveExpr          F   P.Value
#> gene6  -0.67049689 -1.034776107 -1.04729128  -3.1554790 0.49876352 0.6843109
#> gene8   0.02567263 -0.129563777  0.67499421   0.9171386 0.36318893 0.7797634
#> gene3   0.08200976 -0.172582199 -0.49673001   0.1702122 0.20303106 0.8939750
#> gene7  -0.31606553  0.149452306 -0.03562447  -4.3799381 0.15424030 0.9266352
#> gene1  -0.46092425 -0.453624935 -0.05771812 -10.4664549 0.14981701 0.9294858
#> gene2  -0.49420745 -0.382210977 -0.40013018  -1.1281610 0.11010298 0.9539356
#> gene5   0.26795357  0.006060832  0.27368297   4.7187640 0.06850536 0.9765104
#> gene9  -0.35301416 -0.096570091 -0.12646906  -2.3079042 0.06256423 0.9793859
#> gene10  0.11578714 -0.098856729  0.14565869  -2.3673775 0.04373560 0.9877395
#> gene4   0.11531023  0.122855760  0.13721184  -4.5359748 0.01288501 0.9979822
#>        adj.P.Val      F.std
#> gene6  0.9979822 0.49705501
#> gene8  0.9979822 0.36295415
#> gene3  0.9979822 0.20357178
#> gene7  0.9979822 0.15480749
#> gene1  0.9979822 0.15038173
#> gene2  0.9979822 0.11060913
#> gene5  0.9979822 0.06887981
#> gene9  0.9979822 0.06291398
#> gene10 0.9979822 0.04399733
#> gene4  0.9979822 0.01297045